pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "blaxel-agent"
version = "0.1.0"
description = "Hello World agent: Anthropic Claude via Google ADK with SSE streaming"
requires-python = ">=3.11"
dependencies = [
    "google-adk[extensions]>=0.2.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sse-starlette>=1.6.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.20.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]
